requirements.txt
streamlit>=1.35.0
tensorflow==2.19.0
pillow==10.1.0
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
huggingface_hub>=0.20.0